+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### 1.3.0-alpha.9 (2022-05-12)
 
-- Add BarWidget/BarWidgetUI [#229][https://github.com/CartoDB/carto-react/pull/229]
+- Add BarWidget/BarWidgetUI [#229](https://github.com/CartoDB/carto-react/pull/229)
 - Bump deck.gl 8.7.9 [#405](https://github.com/CartoDB/carto-react/pull/405)
 - Fix filters with null as upper or lower bound [#403](https://github.com/CartoDB/carto-react/pull/403)
 - Use static bins in HistogramWidget [#402](https://github.com/CartoDB/carto-react/pull/402)