+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- Fixes in HistogramWidgetUI [#407](https://github.com/CartoDB/carto-react/pull/407)
+
 ## 1.3
 
 ### 1.3.0-alpha.9 (2022-05-12)