+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Not released
 
+## 1.3
+
+### 1.3.0-alpha.10 (2022-05-12)
+
 - Fixes in HistogramWidgetUI [#407](https://github.com/CartoDB/carto-react/pull/407)
 - Display ellipsis and tooltip for long labels in Legend [#408](https://github.com/CartoDB/carto-react/pull/408)
-
-## 1.3
 
 ### 1.3.0-alpha.9 (2022-05-12)